Clean startup window position
We want to load all our config files before love startup so we can position the window correctly the first time, rather than have it jump.

Move around the code so that conf.lua can already have the values it needs.
