+++ globals.lua
@@ -1,7 +1,6 @@
 require("consts")
 require("queue")
 require("server_queue")
-require("sound_util")
 
 -- keyboard assignment vars
 keys = {}
@@ -55,57 +54,5 @@ small_font = -3 -- small font base-3
 
 default_input_repeat_delay = 20
 default_portrait_darkness = 70
-
-zero_sound = load_sound_from_supported_extensions("zero_music")
-
-  -- Default configuration values
-config = {
-	-- The lastly used version
-	version                       = VERSION,
-
-	 -- Lang used for localization
-	language_code                 = "EN",
-
-	theme                         = default_theme_dir,
-	panels                     	  = default_panels_dir,
-	character                     = random_character_special_value,
-	stage                         = random_stage_special_value,
-
-	ranked                        = true,
-
-	vsync                         = false,
-
-	use_music_from                = "either",
-	-- Level (2P modes / 1P vs yourself mode)
-	level                         = 5,
-	endless_speed                 = 1,
-	endless_difficulty            = 1,
-	-- Player name
-	name                          = "defaultname",
-	-- Volume settings
-	master_volume                 = 100,
-	SFX_volume                    = 100,
-	music_volume                  = 100,
-	-- Debug mode flag
-	debug_mode                    = false,
-	-- Show FPS in the top-left corner of the screen
-	show_fps                      = false,
-	-- Show ingame infos while playing the game
-	show_ingame_infos             = true,
-	-- Enable ready countdown flag
-	ready_countdown_1P            = true,
-	-- Change danger music back later flag
-	danger_music_changeback_delay = false,
-  	input_repeat_delay            = default_input_repeat_delay,
-	-- analytics
-	enable_analytics              = true,
-	-- Save replays setting
-	save_replays_publicly         = "with my name",
-	portrait_darkness             = default_portrait_darkness,
-	popfx                         = true,
-	cardfx_scale                  = 100,
-	renderTelegraph               = true,
-	renderAttacks                 = true
-}
 
 current_use_music_from = "stage" -- either "stage" or "characters", no other values!
